Moved EngineCore and run_engine to core.rs
